+++ fxmanifest.lua
@@ -3,7 +3,7 @@ game 'gta5'
 
 author 'Inkwell'
 description 'spacedb - high-performance database bridge for FiveM'
-version '0.2.7'
+version '0.2.8'
 
 server_scripts {
     'server/lua/spacedb.lua',